Fine_tuning_TrOCR/requirements.txt
pandas
torch
transformers
torchvision
datasets
Pillow

